- automatically reduce the font size of buttons if the first word is long or the total text length is long - more code cleanup
- up the limit for the deviceID to 999999
- Force getting a new ID when the intro is run (which is also run on a reset)
- Save the current ID and restore it when import the preferences
- split the TTS code and types into their own classes
- Added the DCC-EX toolbar button the the Routes and Turnouts/Points screens
- General cleanup and linting of the dcc_ex code
- Visual improvements to the 'about' message on the About screen and log

+++ EngineDriver/src/main/java/jmri/enginedriver/comms/comm_thread.java
@@ -1583,7 +1583,8 @@ public class comm_thread extends Thread {
                         int fnState;
                         for (int i = 0; i < threaded_application.MAX_FUNCTIONS; i++) {
                             try {
-                                fnState = mainapp.bitExtracted(Integer.parseInt(args[4]), 1, i + 1);
+//                                fnState = mainapp.bitExtracted(Integer.parseInt(args[4]), 1, i + 1);
+                                fnState = mainapp.bitExtracted(Long.parseLong(args[4]), 1, i + 1);
                                 if (i==0) Log.d("Engine_Driver", "processDccexLocos(): function:" + i + " state: " + fnState);
                                 processFunctionState(whichThrottle, i, (fnState != 0));
                                 responseStr = "M" + mainapp.throttleIntToString(whichThrottle) + "A" + addr_str + "<;>F" + fnState + (i);
